requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
django-cors-headers==3.7.0
django-mptt==0.11.0
pillow==8.1.0
psycopg2==2.8.6
